Moved io to hal folder, created paging constants
