+++ src/kernel/boot/idt/pic/programmable_interrupt_controller.cpp
@@ -17,7 +17,7 @@
 
 #include "boot/idt/pic/programmable_interrupt_controller.h"
 
-#include "boot/io/io.h"
+#include "hal/io/io.h"
 
 namespace kernel::boot::idt {
